requirements.txt
google-genai
cloudpathlib
numpy
pandas
pydantic
tqdm
statsmodels
scipy
matplotlib
seaborn
nest-asyncio
pulp
